+++ test/easy_localization_context_test.dart
@@ -1,5 +1,8 @@
+import 'dart:io';
+
 import 'package:easy_localization/easy_localization.dart';
 import 'package:easy_localization/src/localization.dart';
+import 'package:easy_logger/easy_logger.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_test/flutter_test.dart';
@@ -36,6 +39,11 @@ class MyWidget extends StatelessWidget {
 
 void main() async {
   SharedPreferences.setMockInitialValues({});
+  EasyLocalization.logger.enableLevels = <LevelMessages>[
+    LevelMessages.error,
+    LevelMessages.warning,
+  ];
+
   await EasyLocalization.ensureInitialized();
   group('BuildContext', () {
     testWidgets(
@@ -251,7 +259,7 @@ void main() async {
             await tester.pumpAndSettle();
 
             expect(_context.locale, Locale('ar', 'DZ'));
-            _context.deleteSaveLocale();
+            await _context.deleteSaveLocale();
           });
         },
       );
@@ -259,7 +267,7 @@ void main() async {
       testWidgets(
         '[EasyLocalization] after deleteSaveLocale test',
         (WidgetTester tester) async {
-          await tester.runAsync(() async {            
+          await tester.runAsync(() async {
             await tester.pumpWidget(EasyLocalization(
               child: MyApp(),
               path: 'i18n',
@@ -273,6 +281,53 @@ void main() async {
             // The async delegator load will require build on the next frame. Thus, pump
             await tester.pumpAndSettle();
 
+            expect(_context.locale, Locale('en', 'US'));
+          });
+        },
+      );
+
+      testWidgets(
+        '[EasyLocalization] device locale  test',
+        (WidgetTester tester) async {
+          await tester.runAsync(() async {
+            await tester.pumpWidget(EasyLocalization(
+              child: MyApp(),
+              path: 'i18n',
+              supportedLocales: [
+                Locale('en', 'US'),
+                Locale('ar', 'DZ')
+              ], // Locale('en', 'US'), Locale('ar','DZ')
+            ));
+            await tester.idle();
+            // The async delegator load will require build on the next frame. Thus, pump
+            await tester.pumpAndSettle();
+
+            expect(_context.deviceLocale.toString(), Platform.localeName);
+          });
+        },
+      );
+
+      testWidgets(
+        '[EasyLocalization] reset device locale  test',
+        (WidgetTester tester) async {
+          await tester.runAsync(() async {
+            await tester.pumpWidget(EasyLocalization(
+              child: MyApp(),
+              path: 'i18n',
+              supportedLocales: [
+                Locale('en', 'US'),
+                Locale('ar', 'DZ')
+              ], // Locale('en', 'US'), Locale('ar','DZ')
+              startLocale: Locale('ar', 'DZ'),
+            ));
+            await tester.idle();
+            // The async delegator load will require build on the next frame. Thus, pump
+            await tester.pumpAndSettle();
+
+            expect(_context.locale, Locale('ar', 'DZ'));
+            // reset to device locale
+            await _context.resetLocale();
+            await tester.pumpAndSettle();
             expect(_context.locale, Locale('en', 'US'));
           });
         },